Move view state into server view state module
